refactor: moved icons to icon folder
