--- apps/web/src/components/workflow/Preview/common/icons/android/AndroidBattery.tsx
@@ -1,12 +1,0 @@
-export function AndroidBattery() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="8" height="13" fill="none" viewBox="0 0 8 13">
-      <path
-        fill="#1E1E26"
-        fillRule="evenodd"
-        d="M3.16 0a1 1 0 00-1 1h-.947a1 1 0 00-1 1v10a1 1 0 001 1H7a1 1 0 001-1V2a1 1 0 00-1-1h-.947a1 1 0 00-1-1H3.16z"
-        clipRule="evenodd"
-      ></path>
-    </svg>
-  );
-}
