+++ apps/web/src/components/workflow/Preview/common/icons/android/AndroidIndicatorsIcon.tsx
