--- apps/web/src/components/workflow/Preview/common/icons/android/AndroidNetwork.tsx
@@ -1,12 +1,0 @@
-/* eslint-disable max-len */
-
-export function AndroidNetwork() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="13" height="13" fill="none" viewBox="0 0 13 13">
-      <path
-        fill="#1E1E26"
-        d="M12.665 12V1.461c0-.897-1.09-1.34-1.717-.698L.69 11.303C.072 11.936.521 13 1.406 13h10.259a1 1 0 001-1z"
-      ></path>
-    </svg>
-  );
-}
