--- apps/web/src/components/workflow/Preview/common/icons/android/AndroidWifi.tsx
@@ -1,12 +1,0 @@
-/* eslint-disable max-len */
-
-export function AndroidWifi() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="16" height="13" fill="none" viewBox="0 0 16 13">
-      <path
-        fill="#1E1E26"
-        d="M15.61 4.477c.343-.38.346-.962-.023-1.316A11.574 11.574 0 0012.417.99 10.695 10.695 0 007.938 0C6.402 0 4.88.336 3.46.99A11.574 11.574 0 00.29 3.16c-.37.354-.366.935-.024 1.316L7.939 13l7.671-8.523z"
-      ></path>
-    </svg>
-  );
-}
