+++ apps/web/src/components/workflow/Preview/common/icons/android/index.ts
@@ -1,3 +1,3 @@
-export * from './AndroidNetwork';
-export * from './AndroidWifi';
-export * from './AndroidBattery';
+export * from './AndroidIcon';
+export * from './AndroidIndicatorsIcon';
+export * from './AndroidKeyboard';
