--- apps/web/src/components/workflow/Preview/common/icons/iphone/IphoneBattery.tsx
@@ -1,9 +1,0 @@
-export function IphoneBattery() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="24" height="12" fill="none" viewBox="0 0 24 12">
-      <rect width="20.703" height="11" x="0.5" y="0.5" stroke="#1E1E26" rx="3.5"></rect>
-      <path fill="#1E1E26" d="M22.69 3.882v4.236C23.484 7.758 24 6.924 24 6c0-.925-.516-1.759-1.31-2.118z"></path>
-      <rect width="17.757" height="8.471" x="1.973" y="1.765" fill="#1E1E26" rx="2"></rect>
-    </svg>
-  );
-}
