--- apps/web/src/components/workflow/Preview/common/icons/iphone/IphoneCameraNotch.tsx
@@ -1,12 +1,0 @@
-/* eslint-disable max-len */
-
-export function IphoneCameraNotch() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="160" height="20" fill="none" viewBox="0 0 160 20">
-      <path
-        fill="#1E1E26"
-        d="M0 0h160a3.587 3.587 0 00-2.678 3.34l-.013.368V4c0 8.837-7.164 16-16 16H18.691c-8.836 0-16-7.163-16-16v-.292l-.013-.368A3.586 3.586 0 000 0z"
-      ></path>
-    </svg>
-  );
-}
