--- apps/web/src/components/workflow/Preview/common/icons/iphone/IphoneNetwork.tsx
@@ -1,14 +1,0 @@
-/* eslint-disable max-len */
-
-export function IphoneNetwork() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="18" height="12" fill="none" viewBox="0 0 18 12">
-      <path
-        fill="#000"
-        fillRule="evenodd"
-        d="M16.941 0h-1.059c-.584 0-1.059.504-1.059 1.125v9.75c0 .621.475 1.125 1.06 1.125h1.058c.585 0 1.059-.504 1.059-1.125v-9.75C18 .504 17.526 0 16.941 0zm-6 2.625H12c.585 0 1.059.504 1.059 1.125v7.125c0 .621-.474 1.125-1.059 1.125h-1.059c-.585 0-1.059-.504-1.059-1.125V3.75c0-.621.474-1.125 1.06-1.125zM7.06 5.25H6c-.585 0-1.059.504-1.059 1.125v4.5C4.941 11.496 5.415 12 6 12h1.059c.585 0 1.059-.504 1.059-1.125v-4.5c0-.621-.474-1.125-1.06-1.125zM2.118 7.5h-1.06C.475 7.5 0 8.004 0 8.625v2.25C0 11.496.474 12 1.059 12h1.059c.584 0 1.058-.504 1.058-1.125v-2.25c0-.621-.474-1.125-1.058-1.125z"
-        clipRule="evenodd"
-      ></path>
-    </svg>
-  );
-}
