--- apps/web/src/components/workflow/Preview/common/icons/iphone/IphoneWifi.tsx
@@ -1,14 +1,0 @@
-/* eslint-disable max-len */
-
-export function IphoneWifi() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="16" height="12" fill="none" viewBox="0 0 16 12">
-      <path
-        fill="#000"
-        fillRule="evenodd"
-        d="M8 2.492c2.322 0 4.555.932 6.237 2.603.127.129.33.127.454-.004l1.211-1.277a.36.36 0 00-.003-.499c-4.416-4.42-11.382-4.42-15.798 0a.36.36 0 00-.003.499l1.211 1.277a.313.313 0 00.454.004C3.446 3.424 5.68 2.492 8 2.492zm0 4.154a5.02 5.02 0 013.451 1.39c.128.126.33.124.454-.007l1.21-1.276a.361.361 0 00-.005-.503c-2.879-2.797-7.338-2.797-10.217 0a.36.36 0 00-.005.502l1.21 1.277c.125.13.326.133.454.007A5.02 5.02 0 018 6.646zm2.424 2.795a.351.351 0 01-.099.252L8.233 11.9A.32.32 0 018 12a.32.32 0 01-.232-.101L5.675 9.693a.352.352 0 01.01-.5 3.469 3.469 0 014.63 0 .35.35 0 01.109.248z"
-        clipRule="evenodd"
-      ></path>
-    </svg>
-  );
-}
