+++ apps/web/src/components/workflow/Preview/common/icons/iphone/index.ts
@@ -1,4 +1,4 @@
-export * from './IphoneNetwork';
-export * from './IphoneCameraNotch';
-export * from './IphoneWifi';
-export * from './IphoneBattery';
+export * from './AppleIcon';
+export * from './IOSIndicatorsIcon';
+export * from './IOSKeyboard';
+export * from './TimeIcon';
